Breakout TrustAnchor logic from AuthenticatorAttestationResponse
